Speed up filter feature selection
scikit-learn's SelectKBest computes filter scores with each fit(), while
the number of features is already passed when creating the object. This is
inefficient if one wants to try multiple k (= numbers of features) for the
same dataset. Thus, we replace SelectKBest with our own solution, which
separates score computation and feature selection.