Make database lookup more robust
When merging the GBD databases, the code looked for all CSVs in one
directory. This is error-prone, as not only the databases might be stored
in the directory, but also other files, particularly if we save the merged
dataset and experimental results to the same directory as well (which we do
since 9b2bbec880). As a fix, we directly iterate over the database names.