Store SHAP importances in pipeline
After model-based importance (78141cf3) and filter importance (51229b8),
this is the third feature-importance measure we use for evaluation.
Measures local importance, but can be aggregated (using "SAGE" for global
importance was very slow). Significantly increases pipeline runtime
(from about 2 min to about 30 min on server), but overall runtime is not
a problem yet (if it becomes one, we could compute SHAP only for decision
trees and not for ensembles, or change the parametrization of SHAP).